refactor: retire startup service components
