+++ rustfs/src/lib.rs
@@ -68,10 +68,13 @@ pub mod profiling;
 pub mod protocols;
 pub mod runtime_capabilities;
 pub mod server;
+pub mod startup_audit;
 pub mod startup_auth;
 pub mod startup_background;
 pub mod startup_bucket_metadata;
+pub mod startup_deadlock;
 pub(crate) mod startup_embedded;
+pub mod startup_embedded_optional;
 pub mod startup_entrypoint;
 pub mod startup_fs_guard;
 pub mod startup_iam;
@@ -86,7 +89,6 @@ pub mod startup_protocols;
 pub mod startup_runtime;
 pub mod startup_runtime_hooks;
 pub mod startup_server;
-pub mod startup_service_components;
 pub mod startup_services;
 pub mod startup_shutdown;
 pub mod startup_storage;
