+++ rustfs/src/startup_audit.rs
@@ -12,22 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    config::Config,
-    init::{init_buffer_profile_system, init_kms_system},
-    server::{init_event_notifier, start_audit_system},
-};
+use crate::server::{init_event_notifier, start_audit_system};
 use rustfs_audit::AuditResult;
 use std::future::Future;
-use tracing::{error, info, warn};
+use tracing::{error, info};
 
 const LOG_COMPONENT_MAIN: &str = "main";
-const LOG_COMPONENT_EMBEDDED: &str = "embedded";
 const LOG_SUBSYSTEM_STARTUP: &str = "startup";
-const LOG_SUBSYSTEM_EMBEDDED: &str = "embedded";
 const EVENT_AUDIT_SYSTEM_STATE: &str = "audit_system_state";
-const EVENT_DEADLOCK_DETECTOR_STATE: &str = "deadlock_detector_state";
-const EVENT_EMBEDDED_OPTIONAL_SERVICE_SKIPPED: &str = "embedded_optional_service_skipped";
 
 pub(crate) async fn init_audit_runtime() {
     match init_event_notifier_and_audit().await {
@@ -51,42 +43,6 @@ pub(crate) async fn init_audit_runtime() {
     }
 }
 
-pub(crate) async fn init_embedded_optional_service_runtime(config: &Config) {
-    if let Err(err) = init_kms_system(config).await {
-        log_embedded_optional_service_skipped("kms", err);
-    }
-
-    init_buffer_profile_system(config);
-
-    if let Err(err) = init_event_notifier_and_audit().await {
-        log_embedded_optional_service_skipped("audit", err);
-    }
-}
-
-pub(crate) fn init_deadlock_detector_runtime() {
-    let detector = crate::storage::deadlock_detector::get_deadlock_detector();
-    if detector.is_enabled() {
-        detector.start();
-        info!(
-            target: "rustfs::main::run",
-            event = EVENT_DEADLOCK_DETECTOR_STATE,
-            component = LOG_COMPONENT_MAIN,
-            subsystem = LOG_SUBSYSTEM_STARTUP,
-            state = "started",
-            "Deadlock detector started"
-        );
-    } else {
-        info!(
-            target: "rustfs::main::run",
-            event = EVENT_DEADLOCK_DETECTOR_STATE,
-            component = LOG_COMPONENT_MAIN,
-            subsystem = LOG_SUBSYSTEM_STARTUP,
-            state = "disabled",
-            "Deadlock detector disabled"
-        );
-    }
-}
-
 pub(crate) async fn init_event_notifier_and_audit() -> AuditResult<()> {
     init_event_notifier_and_audit_with(init_event_notifier, start_audit_system).await
 }
@@ -103,17 +59,6 @@ where
 {
     notify().await;
     start_audit().await
-}
-
-fn log_embedded_optional_service_skipped(service: &str, err: impl std::fmt::Display) {
-    warn!(
-        component = LOG_COMPONENT_EMBEDDED,
-        subsystem = LOG_SUBSYSTEM_EMBEDDED,
-        event = EVENT_EMBEDDED_OPTIONAL_SERVICE_SKIPPED,
-        service,
-        error = %err,
-        "Embedded optional service initialization skipped"
-    );
 }
 
 #[cfg(test)]
